extract landmark bearing sensors into separate file
